Limit target voltage to max voltage and fix use of max voltage in voltage measurement calibration.

+++ include/tDCDC.h
@@ -25,7 +25,7 @@ public:
 	double get_Ki();
 	double get_Kd();
 
-	uint16_t get_last_Vcurrent();
+	double get_last_Vcurrent();
 	uint16_t get_last_Vnow();
 	uint16_t get_Vset();
 	uint16_t get_Vmax();
@@ -66,7 +66,7 @@ private:
 	double C1;
 	double C2;
 	uint16_t last_Vnow;
-	uint16_t last_Vcurrent;
+	double last_Vcurrent;
 
 	double get_HV_voltage(uint8_t nAvg); //deprecated
 	double get_HV_voltage_fast(float alpha);
